Add test for runN cli use
